requirements.txt
django==3.1.6
djangorestframework==3.12.2
django_phonenumber_field==5.0.0
uuid==1.30
mysqlclient==2.0.3
phonenumbers==8.12.18

